TP3/devfotografia/requirements.txt
django
djangorestframework
djangorestframework-jwt
psycopg2-binary
requests
django-allauth
dj-rest-auth[with_social]==4.0.0